# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PreviewShield API - Secure file preview sharing platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreatePreviewRequestFileType =
  (typeof CreatePreviewRequestFileType)[keyof typeof CreatePreviewRequestFileType];

export const CreatePreviewRequestFileType = {
  image: "image",
  pdf: "pdf",
  video: "video",
} as const;

export interface CreatePreviewRequest {
  freelancerName: string;
  agencyName?: string | null;
  clientName: string;
  fileName: string;
  fileType: CreatePreviewRequestFileType;
  fileMimeType: string;
  fileSize: number;
  password?: string | null;
  fileUrl?: string | null;
}

export interface Preview {
  id: string;
  freelancerName: string;
  agencyName?: string | null;
  clientName: string;
  fileName: string;
  fileType: string;
  fileMimeType: string;
  fileSize: number;
  fileUrl?: string | null;
  hasPassword: boolean;
  previewUrl: string;
  ownerToken: string;
  createdAt: string;
  expiresAt?: string | null;
}

export interface PreviewPublic {
  id: string;
  freelancerName: string;
  agencyName?: string | null;
  clientName: string;
  fileName: string;
  fileType: string;
  fileMimeType: string;
  fileSize: number;
  fileUrl?: string | null;
  hasPassword: boolean;
  createdAt: string;
}

export interface RecordVisitRequest {
  userAgent?: string | null;
  referrer?: string | null;
}

export interface VisitRecorded {
  success: boolean;
  visitId: string;
}

export interface Visit {
  id: string;
  previewId: string;
  ipAddress?: string | null;
  userAgent?: string | null;
  referrer?: string | null;
  visitedAt: string;
}

export interface VisitsList {
  visits: Visit[];
  total: number;
}

export interface UploadUrlRequest {
  fileName: string;
  fileMimeType: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  fileUrl: string;
}

export interface PreviewStats {
  previewId: string;
  totalVisits: number;
  uniqueIps: number;
  lastVisitAt?: string | null;
  recentVisits: Visit[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetPreviewParams = {
  password?: string;
};

export type GetPreviewVisitsParams = {
  ownerToken: string;
};

export type GetPreviewStatsParams = {
  ownerToken: string;
};