# lib/api-zod/src/generated/types/createPreviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PreviewShield API - Secure file preview sharing platform
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePreviewRequestFileType } from "./createPreviewRequestFileType";

export interface CreatePreviewRequest {
  freelancerName: string;
  agencyName?: string | null;
  clientName: string;
  fileName: string;
  fileType: CreatePreviewRequestFileType;
  fileMimeType: string;
  fileSize: number;
  password?: string | null;
  fileUrl?: string | null;
}