# lib/api-zod/src/generated/types/preview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PreviewShield API - Secure file preview sharing platform
 * OpenAPI spec version: 0.1.0
 */

export interface Preview {
  id: string;
  freelancerName: string;
  agencyName?: string | null;
  clientName: string;
  fileName: string;
  fileType: string;
  fileMimeType: string;
  fileSize: number;
  fileUrl?: string | null;
  hasPassword: boolean;
  previewUrl: string;
  ownerToken: string;
  createdAt: Date;
  expiresAt?: Date | null;
}