# lib/api-zod/src/generated/types/previewPublic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PreviewShield API - Secure file preview sharing platform
 * OpenAPI spec version: 0.1.0
 */

export interface PreviewPublic {
  id: string;
  freelancerName: string;
  agencyName?: string | null;
  clientName: string;
  fileName: string;
  fileType: string;
  fileMimeType: string;
  fileSize: number;
  fileUrl?: string | null;
  hasPassword: boolean;
  createdAt: Date;
}