# lib/api-zod/src/generated/types/previewStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PreviewShield API - Secure file preview sharing platform
 * OpenAPI spec version: 0.1.0
 */
import type { Visit } from "./visit";

export interface PreviewStats {
  previewId: string;
  totalVisits: number;
  uniqueIps: number;
  lastVisitAt?: Date | null;
  recentVisits: Visit[];
}